Move measurements API to API app
