+++ api/views.py
@@ -1,3 +1,12 @@
-from django.shortcuts import render
+from rest_framework import viewsets
+from django.views   import generic
+from .models        import Measurement
+from .serializers   import MeasurementSerializer
 
-# Create your views here.
+
+class MeasurementViewset(viewsets.ModelViewSet):
+    # define queryset
+    queryset = Measurement.objects.all()
+ 
+    # specify serializer to be used
+    serializer_class = MeasurementSerializer
