+++ sensor_data/views.py
@@ -1,7 +1,6 @@
 from rest_framework import viewsets
 from django.views   import generic
 from .models        import Measurement, Sensor
-from .serializers   import MeasurementSerializer
 
 # ==========================================
 # Root of app. Fetch a list of all devices
@@ -52,12 +51,3 @@ class SensorView(generic.ListView):
 
 def getSensors():
     return Sensor.objects.order_by('name')
-
-
-
-class MeasurementViewset(viewsets.ModelViewSet):
-    # define queryset
-    queryset = Measurement.objects.all()
- 
-    # specify serializer to be used
-    serializer_class = MeasurementSerializer
